Create entity id counter used to generate ids (#63)
Move entity ids to u64
Add module address, host and ssl options on BitcraftMiniGameManager inspector

+++ Client/Assets/_Project/Chat/UIChatController.cs
@@ -71,7 +71,7 @@ public class UIChatController : Singleton<UIChatController>
         Canvas.ForceUpdateCanvases();
     }
     
-    public void OnChatMessageReceived(uint playerId, String message)
+    public void OnChatMessageReceived(ulong playerId, String message)
     {
         var player = PlayerComponent.FilterByEntityId(playerId);
         if (player != null)
