+++ Client/Assets/_Project/Inventory/UIPlayerInventoryWindow.cs
@@ -17,7 +17,7 @@ public class UIPlayerInventoryWindow : UIInventoryWindow
         instance = null;
     }
 
-	protected override void CallReducer(uint playerEntityId, UIInventorySlot source, UIInventorySlot dest)
+	protected override void CallReducer(ulong playerEntityId, UIInventorySlot source, UIInventorySlot dest)
 	{
         if (source.Owner == this)
         {
