+++ Server/src/components/animation_component.rs
@@ -3,7 +3,7 @@ use spacetimedb::spacetimedb;
 #[spacetimedb(table)]
 pub struct AnimationComponent {
     #[unique]
-    pub entity_id: u32,
+    pub entity_id: u64,
     pub moving: bool,
-    pub action_target_entity_id: u32,
+    pub action_target_entity_id: u64,
 }
