+++ Server/src/components/npc_component.rs
@@ -3,7 +3,7 @@ use spacetimedb::spacetimedb;
 #[spacetimedb(table)]
 pub struct NpcComponent {
     #[unique]
-    pub entity_id: u32,
+    pub entity_id: u64,
     pub model: String,
     pub next_action: u64,
 }
