+++ Server/src/components/player_login_component.rs
@@ -3,6 +3,6 @@ use spacetimedb::spacetimedb;
 #[spacetimedb(table)]
 pub struct PlayerLoginComponent {
     #[unique]
-    pub entity_id: u32,
+    pub entity_id: u64,
     pub logged_in: bool,
 }
