+++ Server/src/components/resource_component.rs
@@ -3,7 +3,7 @@ use spacetimedb::spacetimedb;
 #[spacetimedb(table)]
 pub struct ResourceComponent {
     #[unique]
-    pub entity_id: u32,
+    pub entity_id: u64,
     pub health: u8,
     pub resource_id: u8,
     pub max_health: u8, // todo: ideally we would find that static data from a table using resource_id
