+++ Server/src/components/trade_session_component.rs
@@ -4,11 +4,11 @@ use spacetimedb::spacetimedb;
 #[derive(Clone)]
 pub struct TradeSessionComponent {
     #[unique]
-    pub entity_id: u32,
-    pub initiator_entity_id: u32,
-    pub acceptor_entity_id: u32,
-    pub acceptor_offer_inventory_entity_id: u32,
-    pub initiator_offer_inventory_entity_id: u32,
+    pub entity_id: u64,
+    pub initiator_entity_id: u64,
+    pub acceptor_entity_id: u64,
+    pub acceptor_offer_inventory_entity_id: u64,
+    pub initiator_offer_inventory_entity_id: u64,
     pub approved_by_initiator: bool,
     pub approved_by_acceptor: bool,
 }
@@ -16,6 +16,6 @@ pub struct TradeSessionComponent {
 #[spacetimedb(table)]
 pub struct ActiveTradeComponent {
     #[unique]
-    pub entity_id: u32,
-    pub trade_session_entity_id: u32,
+    pub entity_id: u64,
+    pub trade_session_entity_id: u64,
 }
