+++ Server/src/npcs/npc_reducers.rs
@@ -6,7 +6,7 @@ use spacetimedb::hash::Hash;
 use spacetimedb::spacetimedb;
 
 #[spacetimedb(reducer)]
-pub fn move_npc(_identity: Hash, timestamp: u64, entity_id: u32, pos: StdbVector3, rot: StdbQuaternion, duration: u64) {
+pub fn move_npc(_identity: Hash, timestamp: u64, entity_id: u64, pos: StdbVector3, rot: StdbQuaternion, duration: u64) {
     /*
     TODO: Uncomment when supported.
     if identity != 0 {
@@ -26,9 +26,9 @@ pub fn move_npc(_identity: Hash, timestamp: u64, entity_id: u32, pos: StdbVector
 pub fn update_npc_animation(
     _identity: Hash,
     _timestamp: u64,
-    entity_id: u32,
+    entity_id: u64,
     moving: bool,
-    action_target_entity_id: u32,
+    action_target_entity_id: u64,
 ) {
     let _npc = NpcComponent::filter_by_entity_id(entity_id).expect("This npc doesn't exist.");
